DC-1757: attempting to fix IT tests
